ENH: delegate sinc function.
